split out IDGenerator
