+++ profile/src/wallet_kit_common/mod.rs
@@ -2,6 +2,7 @@ mod assert_json;
 mod error;
 mod has_placeholder;
 mod hash;
+mod id_generator;
 mod secure_random_bytes;
 mod types;
 mod utils;
@@ -10,6 +11,7 @@ pub use assert_json::*;
 pub use error::*;
 pub use has_placeholder::*;
 pub use hash::*;
+pub use id_generator::*;
 pub use secure_random_bytes::*;
 pub use types::*;
 pub use utils::*;
